Add noop migration to force tests load db schema
Rails 4.1 instroduces some intelligence about migrations and schema loading
for databases in test envs. Because we're doing something unorthodox here,
it needs to be told that there is a migration pending for the test database
(i.e. migration 1, which is a noop) before it will attempt to load the schema.