Remove `length` property in favor of `duration`.
In JSON, consider `length` a reserved word which can only be used to
refer to the number of iterable keys in an array-like object.

https://github.com/angular/angular.js/blob/master/src/Angular.js#L185
The current behavior causes stories with a zero duration to have no
properties copied from the JSON response because Angular detects them
as array-like and having no properties.

Also bumps to a newer version of guard-minitest.