Simplify license display to reflect plans for license terms
Right now we have two sets of radio buttons that really should just be
two checkboxes. Specifically:

Allow Edit: "only with permission" and "never" seem to essentially mean the
same thing, so we collapse them.

Allow use on website: downloading as an MP3 doesn't really seem to be a thing
stations care about offering, and if they do they are always welcome to ask.
Thus, we can simplify the choice to whether or not streaming on a licensee's
website is ok.