Make `zoom` behavior somewhat clearer, provide sane default.
Clean up embeds test to make the documented behavior clearer to me. I *think* I
translated from what was there to the new tests with fidelity, but I may have
misunderstood.

Switch the default case from `false` (which is effectively a noop until zooms
start being requested) to `true`, which was the value used in almost every case.

Bring back behavior which conditionally nests depending on the depth of the
structure so far, which was documented but looks like it was removed at some
point. This makes `true` and `:always` no longer equivalent.

remove `zoom: true` from all embeds using it (most of them), and add
`zoom: false` to the embeds which we don't actually want.

Intuition tells me that the `:always` and `true` behavior may be the only
behavior we need in the future, which would further simplify/clarify this.