Add support for linking to external websites on accounts
`external` seemed like the appropriate rel to use from the list of existing and
proposed rels on the microformats wiki. I opted not to add it to the list of
link types that would not be prefixed, but I anticipate that we may opt to do so
in the near future, as it's already in common, unprefixed use.

http://microformats.org/wiki/rel-external
http://microformats.org/wiki/existing-rel-values#HTML5_link_type_extensions

The only other rels I could find that might make sense are `alternate` (but I
don't like this for a few reasons) or `me` which might make more sense for
specifically social links (like facebook, twitter, etc).