Automatically mark stories as deleted and `v4`
Mark stories as deleted when created through cms.prx.org so that they do not
appear in v3.

Mark stories as `v4` so that we can show them in cms.prx.org even though they
are marked as deleted.

When a v3 story is deleted through cms.prx.org, mark as deleted.

When a v4 story is deleted through cms.prx.org, remove the record from the
database.

resolves #87