use validation callback for updating attributes
we're picking up some complexity here for the sake of optimizations. practically
speaking rails will usually call `valid?` automatically before it matters, but
obviously this means we have to manually do it in one form or another in our
tests. we also lose the ability to construct objects which contradict these
defaults directly, since we need to make changes after the validate callback.
this is reflected in the tests.

probably worth these tradeoffs, just introducing leakiness into the abstraction.