Trust signed token instead of `users` and `memberships` tables
For the purposes of authorization, rely on the inforamtion encoeded in the
signed token instead of going to the database.

This allows us to start making tokens that do not line up perfectly with what is
stored in the database (i.e. application agents, which are not members of the
accounts but are still granted permissions) and also allows us to issue tokens
that only work on a subset of the permissions a given user actually posesses.

There are a couple of holdouts here. For one, the `authorization` endpoint still
relies on the `memberships` table to show the accounts that the user associated
with the currently active token has access to. We're also being a little fast
and loose with our use of scopes – most checks are for the mere presence of the
account identifier in the token, regardless of scopes granted. Those that do
look for a specific scope are checking for the `admin` scope which is, at best,
a meta-scope which encompasses several more specific ones. Those should be
specified further.