Yak-Shaving: Make `tap` work as expected on Minitest::Mock instances
`tap` is used at various points in the codebase. It's supposed to be
the k-combinator but in the case of a mock it gets passed to the
delegated instance which makes it effectively unmock in the context of
the tap call. This change ensures that tap doesn't accidentally unwrap
a mock to its delegate.