Call `transform!` on image instances after create or update
`transform!` will eventually be the method that sends tasks to Porter.

After discussing with @cavis, we agreed it made sense to have the calls
be relatively more explicit in the controllers rather than either
relying on implicit reactions to actions from `announce` or relying on
model callback methods. It's still a little magical as implemented here
since it's in a base controller class, but hopefully explicit enough to
be easy to parse.

I'm willing to convert the superclass to a concern/mixin, but in this
case my first instinct is that a superclass makes sense given the
relationship between the models being operated on.